[commission_api] Create commission plans

+++ commission/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class Plan(models.Model):
 
-    min_value = models.DecimalField(max_digits=10, decimal_places=2)
+    min_value = models.FloatField()
     lower_percentage = models.FloatField()
     upper_percentage = models.FloatField()
     created_at = models.DateTimeField('Criado em', auto_now_add=True)
